Backend-d/requirements.txt
absl-py==1.3.0
adversarial-robustness-toolbox==1.13.0
appdirs==1.4.4
asgiref==3.6.0
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
flatbuffers==23.1.4
fonttools==4.38.0
frozendict==2.3.4
funcy==1.17
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
html5lib==1.1
idna==3.4
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multitasking==0.0.11
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
# scikit-learn==0.24.2
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
yfinance==0.2.3
